+++ src/python/DAS/utils/utils.py
@@ -1021,7 +1021,19 @@ def xml_parser(source, prim_key, tags=None):
     """
     notations = {}
     sup       = {}
-    context   = ET.iterparse(source, events=("start", "end"))
+    try:
+        context = ET.iterparse(source, events=("start", "end"))
+    except IOError as exc: # given source is not parseable
+        # try different data format, it can be an HTTP error
+        try:
+            if  isinstance(source, str):
+                data = json.loads(source)
+                yield data
+        except:
+            pass
+        msg = 'XML parser, data stream is not parseable: %s' % str(exc)
+        print_exc(msg)
+        context = []
     root      = None
     for item in context:
         event, elem = item
@@ -1052,8 +1064,10 @@ def xml_parser(source, prim_key, tags=None):
         if  event == 'end':
             elem.clear()
             yield row
-    root.clear()
-    source.close()
+    if  root:
+        root.clear()
+    if  isinstance(source, InstanceType) or isinstance(source, file):
+        source.close()
 
 def get_children(elem, event, row, key, notations):
     """